Introduce new attribute: comptime
Applies to traits to mark that their associated items are a template that
must be filled by the source code procued by a compile time executed
function. This implies compiler generated impls, which are to be produced
during selection for Function Item Types (those anonymous zero-sized types
that exist for each function) where the function is `const fn` and has a
fitting signature, and evaluates without panicking, and produces the right
bound.

We will use the original site requesting that trait impl to annotate the
failure of finding an impl with any error or panic that occurs.